refactor: clean up component architecture
