+++ components/sections/projects/big-projects/BigProjectCard-reversed.tsx
