+++ components/sections/projects/big-projects/BigProjectCard.tsx
