+++ components/sections/projects/big-projects/BigProjectCardReversed.tsx
