+++ components/sections/projects/small-projects/SmallProjectCard.tsx
